28_04_2023

+++ Colmado/Data/Producto.cs
@@ -5,5 +5,5 @@ public class Producto
     public string Descripcion {get; set;}
     public float Cantidad {get; set;}
     public decimal Precio {get; set;}
-    public decimal SubTotal => Cantidad * Precio;
+    public decimal SubTotal => (decimal)Cantidad * Precio;
 }
